Remove clients from other stacks
It was thought to be a good idea to have it possible to switch
to any client from any stack via the stack menu, which was
showing all clients. It was technically very well possible.
However, it made it harder to quickly switch to next client
in the current stack.