Modify move client behavior
Now it is easier to move client from one stack to another because
it is no longer necessary to select the client from the menu, but
it is by default assumed the focused client is to be selected.

Previously it felt logically correct to do the action to the stack
when the stack was selected (i.e. the menu was not open), but it
is easier for the user to simply use a different modifier key when
wanting to do action for the stack.

Also fixed a minor bug when moving the client from middle of the
menu.