Assign clients to stacks based on x,y
Do this only initially when starting, not after that because
having clients appear randomly to any stack is not nice.

Doing this when starting is useful because it allows restarting
window manager without having to rearrange clients. It also
allows stack selection for clients when starting when using e.g.
-geometry flags from a startup script (e.g. from .xsession
combined with .mxswmrc).