Support _NET_WM_VISIBLE_NAME
Previously user set client title was overridden by the client
because the _NET_WM_NAME atom was shared for both the client's
requested title and the user set title. Now they're in two
separate properties.

Giving the control back to the client is possible by renaming
without making a change, or by submitting an empty title.